FEAT: add some convenience getters
